Build scripts now detect Leopard and when present use it's Terminal - thanks for the tip Allan.
Added command to fetch the output of the last compile from the Leopard Terminal and format it - linked errors etc - as this is now possible to do. Once I have worked on the applescript a little more I am hoping it will be possible to automate this when running build.

Help updated.

Comment removal in templates.

Menu updated.

git-svn-id: http://svn.textmate.org/trunk/Review/Bundles/ActionScript 3.tmbundle@8474 dfb7d73b-c2ec-0310-8fea-fb051d288c6d