• Adding a command which allows users to insert a tmTemplate via a tab triggered command.
The command searches the bundles template directory for templates containing a
class.as file and presents them in a list for the user to pick from. Additional work will be needed to scan a user modified bundle but this starts to cover
using 'traditional' templates as well as the snippetised variety. Doing this gives the ability to filter asdoc tags and comment banners via env vars.

git-svn-id: http://svn.textmate.org/trunk/Review/Bundles/ActionScript 3.tmbundle@11364 dfb7d73b-c2ec-0310-8fea-fb051d288c6d