• Class Template (tab triggered on cla) now searches across bundles for templates.
By default the templates found in the ActionScript 3 and Flex bundles are included. This is easy to expand by adding the names of alternative bundles you want included to the env var TM_AS3_TEMPLATE_BUNDLES (see Settings.tmPreference). Any template which defines holds a 'class.as' file will be included.