• Compressed and rewritten Build commands into a single command.
This implementation is now written in ruby and should behave in the same way as the previous Build command. Additionally - when running a build using fcsh and the Terminal the Terminal is activated. To suppress this behavior set TM_FLEX_BACKGROUND_TERMINAL to anything.

• 'Intelligent' Run command.

Now inspects TM_FLEX_RUN_FILE to see if it's executable, if it is then it runs it. Otherwise it uses `open` as open should correctly identify and open the file using the users default application (ie Safari for a html file, SA Flash Player for a swf). When TM_FLEX_RUN_FILE isn't available a little educated guesswork is used to try an identify a compiled swf to open that matches the current file name.

• Tooltip help is now available within flex config files (on ctrl-h).

• Spring cleaning for templates and snippets - making sure they're conforming to standards, trimming whitespace etc.

• Auto Completion and Build commands abstracted to the bin to try and avoid path problems when the bundle finds itself in the 'Pristine Copy' directory.

• Help updated to reflect the changes above.

• Added a 'Subclass' drag command.

If you drag another class into the current document you'll be presented with options to 'include', 'implement interface' or 'subclass'.

• Class Template (tab triggered on cla) now searches across bundles for templates.

By default the templates found in the ActionScript 3 and Flex bundles are included. This is easy to expand by adding the names of alternative bundles you want included to the env var TM_AS3_TEMPLATE_BUNDLES (see Settings.tmPreference). Any template which defines holds a 'class.as' file will be included.

• All fcsh/mxmlc output parsers now use an instance of MxmlcExhaust.

• When detecting if a class has already been imported the regex now only matches full names.

Previously partial class names were being picked up, so the UI could report a class was imported when it wasn't. Closes #1

git-svn-id: http://svn.textmate.org/trunk/Review/Bundles/ActionScript 3.tmbundle@11568 dfb7d73b-c2ec-0310-8fea-fb051d288c6d